Move vppapi commands
